+++ client/package.json
@@ -11,6 +11,7 @@
     "react-dom": "^16.4.1",
     "react-router-dom": "^4.3.1",
     "react-scripts": "1.1.4",
+    "unstated": "^2.1.1",
     "xregexp": "^4.2.0"
   },
   "scripts": {
